move components to new structure
