+++ src/sections/Reports/components/DateRange/index.js
@@ -14,7 +14,7 @@ import subMonths from 'date-fns/sub_months';
 import Button from 'semantic-ui-react/dist/commonjs/elements/Button';
 import Input from 'semantic-ui-react/dist/commonjs/elements/Input';
 
-import { valueFromEventTarget } from '../../core/dom';
+import { valueFromEventTarget } from '../../../../core/dom';
 
 import './ReportRange.css';
