+++ src/sections/Reports/components/Detailed/ReportDetailed.css
