+++ src/sections/Reports/components/SavedReports/index.js
@@ -12,7 +12,7 @@ import Header from 'semantic-ui-react/dist/commonjs/elements/Header';
 import Confirm from 'semantic-ui-react/dist/commonjs/addons/Confirm';
 import Popup from 'semantic-ui-react/dist/commonjs/modules/Popup';
 
-import { addReport, removeReport } from '../../actions/reports';
+import { addReport, removeReport } from '../../../../actions/reports';
 
 import { getAllReports } from '../../selectors/reports';
