+++ src/sections/Reports/components/Table/index.js
@@ -4,7 +4,7 @@ import React from 'react';
 
 import Table from 'semantic-ui-react/dist/commonjs/collections/Table';
 
-import { formatDuration } from '../../core/thyme';
+import { formatDuration } from '../../../../core/thyme';
 
 type ReportTableType = {
   projects: Array<projectTreeWithTimeType>,
